sdk/conformance/context/interface.html

<!doctype html>
<html class="a">
 <head>
  <title>AudioContext interface</title>
  <script src="../../lib/testharness.js"></script>
  <script src="../../lib/testharnessreport.js"></script>
  <script src="../../lib/lodash.js"></script>
 </head>
 <body class="a">
  <div id="log"></div>
  <script>

var attributes = [
  "destination",
  "sampleRate",
  "listener",
  "activeSourceCount"
];

var methods = [
  "createBuffer",
  "decodeAudioData",
  "createBufferSource",
  "createMediaElementSource",
  "createMediaStreamSource",
  "createScriptProcessor",
  "createAnalyser",
  "createGain",
  "createDelay",
  "createBiquadFilter",
  "createWaveShape",
  "createPanner",
  "createConvolver",
  "createChannelSplitter",
  "createChannelMerger",
  "createDynamicsCompressor",
  "createOscillator",
  "createWaveTable"
];

function assert_has_property(obj, name, desc) {
  assert_true(undefined != obj[name], desc);
}

function assert_is_method(obj, name, desc) {
  assert_true("function" === typeof obj[name], desc);
}

function assert_defined(obj, desc) {
  assert_true(undefined != obj, desc);
}

var context;
setup(function() {
  context = new AudioContext();
});

_(attributes).forEach(function(attr) {
  test(function() {
    assert_defined(context, "context is defined")
    assert_has_property(context, attr, "should be an attribute");
    assert_readonly(context, attr, "should be read-only");
  }, attr);
});

_(methods).forEach(function(method) {
  test(function() {
    assert_defined(context, "context is defined")
    assert_is_method(context, method, "should be a method");
  }, method);  
});

  </script>
 </body>
</html>
